using single precision floating point format

+++ snn/source/python/network.cpp
@@ -3,9 +3,9 @@
 #define NDARRAY_FROM_DATA(data) \
     np::from_data( \
         data.memptr(), \
-        np::dtype::get_builtin<double>(), \
+        np::dtype::get_builtin<float>(), \
         p::make_tuple(data.n_rows, data.n_cols), \
-        p::make_tuple(sizeof(double), sizeof(double) * data.n_rows), \
+        p::make_tuple(sizeof(float), sizeof(float) * data.n_rows), \
         m_numpy_memory_owner)
 
 #define NDARRAY_MEMBER(member) \
